requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.7
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-celery-beat==2.2.0
django-celerybeat-status==0.0.10
django-ratelimit==3.0.1
django-timezone-field==4.1.2
djangorestframework==3.12.2
flower==0.9.7
humanize==3.3.0
importlib-metadata==3.7.3
kombu==4.6.11
prometheus-client==0.8.0
prompt-toolkit==3.0.17
psycopg2-binary==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
tornado==6.1
typing-extensions==3.7.4.3
vine==1.3.0
wcwidth==0.2.5
zipp==3.4.1
